+++ src/components/Form/FormComboboxMulti.tsx
@@ -2,7 +2,6 @@ import { Combobox } from '@headlessui/react';
 import { FieldValues } from 'react-hook-form';
 import { GenericDataType, dataItemComparator } from '../../common';
 import { ComboboxProps } from './types';
-import { useMultiSelectFormSync } from './useMultiSelectFormSync';
 
 export const ComboboxMulti = <
   DataItem extends GenericDataType,
@@ -11,31 +10,20 @@ export const ComboboxMulti = <
   children,
   className,
   disabled,
-  formValueMode,
   name,
-  options,
   selectedItems,
   setSelectedItems,
-}: ComboboxProps<DataItem, Values>): JSX.Element => {
-  useMultiSelectFormSync({
-    name,
-    options,
-    selectedItems,
-    setSelectedItems,
-    valueMode: formValueMode,
-  });
-  return (
-    <Combobox
-      as="div"
-      by={dataItemComparator}
-      className={className}
-      disabled={disabled}
-      multiple
-      name={name}
-      onChange={(value: DataItem[]) => setSelectedItems(value)}
-      value={selectedItems}
-    >
-      {children}
-    </Combobox>
-  );
-};
+}: ComboboxProps<DataItem, Values>): JSX.Element => (
+  <Combobox
+    as="div"
+    by={dataItemComparator}
+    className={className}
+    disabled={disabled}
+    multiple
+    name={name}
+    onChange={(value: DataItem[]) => setSelectedItems(value)}
+    value={selectedItems}
+  >
+    {children}
+  </Combobox>
+);