+++ src/components/TypeaheadSelect/TypeaheadSelect.tsx
@@ -1,7 +1,7 @@
 import { Combobox } from '@headlessui/react';
 import classNames from 'classnames';
 import { Fragment, KeyboardEvent, KeyboardEventHandler } from 'react';
-import { FieldValues, useFormContext } from 'react-hook-form';
+import { FieldValues, useController, useFormContext } from 'react-hook-form';
 import { Input, Progress } from 'react-daisyui';
 import { GenericDataType, getGroupedDataItems } from '../../common';
 import { useFieldColor, UseTypeaheadQueryOptions } from '../../hooks';
@@ -10,6 +10,7 @@ import { DropdownMenuItem } from '../DropdownMenu';
 import { FormError, FormFieldProps, FormLabel, FormValueMode } from '../Form';
 import { ComboboxMulti } from '../Form/FormComboboxMulti';
 import { ComboboxSingle } from '../Form/FormComboboxSingle';
+import { useSelectFormSync } from '../Form/useSelectFormSync';
 import { useTypeaheadSelect } from './useTypeaheadSelect';
 
 export interface TypeaheadSelectProps<
@@ -54,12 +55,17 @@ export const TypeaheadSelect = <
   showDirty,
 }: TypeaheadSelectProps<DataItem, Values>): JSX.Element => {
   const {
+    fieldState: { error },
+    field: { ref },
+  } = useController({
+    ...controllerProps,
+    name,
+  });
+  const {
     clearSelectedItem,
     dropdownRef,
-    error,
     isLoading,
     query,
-    ref,
     showDropdown,
     searchInputRef,
     selectedItems,
@@ -67,13 +73,20 @@ export const TypeaheadSelect = <
     setSelectedItems,
     setQuery,
   } = useTypeaheadSelect<HTMLUListElement, DataItem, Values>({
-    controllerProps,
     debounceTime,
-    name,
     onDebouncedChange,
     options: optionsArray,
   });
   const options = getGroupedDataItems(optionsArray ?? []);
+  useSelectFormSync({
+    multi,
+    name,
+    onItemSelect: () => multi !== true && setShowDropdown(false),
+    options,
+    selectedItems,
+    setSelectedItems,
+    valueMode: formValueMode,
+  });
   const { setFocus } = useFormContext();
   const fieldColor = useFieldColor(name, showDirty);
   const enableBadges = disableSingleSelectBadge === undefined || multi === true;
@@ -82,11 +95,8 @@ export const TypeaheadSelect = <
     <Component
       className={classNames('relative', className)}
       disabled={disabled}
-      formValueMode={formValueMode}
       name={name}
-      options={options}
       selectedItems={selectedItems}
-      setShowDropdown={setShowDropdown}
       setSelectedItems={items => {
         setSelectedItems(items);
         if (multi === undefined) setFocus(name);